Move XRef iterators to xref module
